finished json stuff
